Intermediate Build
This build is intended to be a jumping-off point to add in the updated board with collision, etc. onto the last screen in the center of a BorderPane.

Updated FXML to include custom backgrounds. Ignore start screen button placement. It's off, I know. Still trying to anchor that to the anchor pane behind it in a specific spot.

Inside the SRC file, I've added a kill graphic that I put together using some transparent free resources I found. We can choose to use that or not. Doesn't matter to me.